float peso Animal

+++ Herencia_Animales/src/herencia_animales/Animal.java
@@ -51,7 +51,7 @@ public abstract class Animal {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(float peso) {
         this.peso = peso;
     }
